+++ src/main/java/ovh/not/javamusicbot/CommandManager.java
@@ -42,6 +42,7 @@ public class CommandManager {
                 new SearchCommand(this),
                 new ShuffleCommand(),
                 new SkipCommand(),
+                new SoundcloudCommand(this, playerManager),
                 new StopCommand(),
                 new VolumeCommand(config)
         );
